refactor: split structured-sidecar types from behavior into homeboy-extension-contract
First Stage-2 cut of the extension campaign (wiki 12840): applies the
data/behavior split pattern to manifest_sidecar.

The three structured-sidecar types (StructuredSidecarContract,
StructuredSidecarDetail, StructuredSidecarDeclaration) are pure serde data, but
the module carried behavior: the declaration() method resolves default paths
and producers via run_dir::files constants and the structured_sidecar module
(both core-only).

Moves the pure types to a new contract sidecar_config module. The resolution
logic becomes a core free function (structured_sidecar_declaration) that keeps
the run_dir/structured_sidecar dependencies. manifest.rs's structured_sidecars()
calls the free function instead of the former inherent method.

This clears the last core-typed field blocking ExtensionManifest itself from
moving to the contract crate (all its other field types are already there).
