+++ crates/homeboy-extension-contract/src/sidecar_config.rs
@@ -1,0 +1,43 @@
+//! Structured-sidecar manifest contract types.
+//!
+//! Pure serde data describing an extension's structured-sidecar declarations.
+//! The logic that resolves a contract into a concrete declaration (default
+//! paths/producers, which depend on core's run-dir file constants) lives in
+//! `homeboy-core` as a free function, so these types stay behavior-free.
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(untagged)]
+pub enum StructuredSidecarContract {
+    Enabled(bool),
+    Detail(StructuredSidecarDetail),
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(deny_unknown_fields)]
+pub struct StructuredSidecarDetail {
+    #[serde(default = "default_true")]
+    pub enabled: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub path: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub schema_version: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub producer: Option<String>,
+}
+
+fn default_true() -> bool {
+    true
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(deny_unknown_fields)]
+pub struct StructuredSidecarDeclaration {
+    pub name: String,
+    pub path: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub schema_version: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub producer: Option<String>,
+}
